+++ examples/lift/lift.hs
@@ -1,35 +1,31 @@
 ------------------------------------------------------------------------------
 -- |
 -- Module      : Main
--- Description : example of compile-time validation failure
+-- Description : example of compile-time validation
 -- Copyright   : Copyright (c) 2019 Travis Cardwell
 -- License     : MIT
 --
--- 'TTC.valid' is used to create validated constants.  The credit card number
--- has an invalid checksum, triggering a compile-time error.
+-- 'TTC.valid' is used to create validated constants.  The sample username
+-- is validated at compile-time.
 ------------------------------------------------------------------------------
 
 {-# LANGUAGE TemplateHaskell #-}
 
 module Main (main) where
 
--- (ttc)
+-- https://hackage.haskell.org/package/ttc
 import qualified Data.TTC as TTC
 
--- (ttc:example-invalid)
-import CreditCard (CreditCard(CreditCard))
+-- (ttc-examples:example-valid)
+import Username (Username)
 
 -- HLint does not support typed expression splices
 {-# ANN module "HLint: ignore" #-}
 
 ------------------------------------------------------------------------------
 
-sample :: CreditCard
-sample = CreditCard
-    $$(TTC.valid "John Q Doe")
-    $$(TTC.valid "1234 5678 9012")
-    $$(TTC.valid "2020-06")
-    $$(TTC.valid "123")
+sample :: Username
+sample = $$(TTC.valid "tcard")
 
 ------------------------------------------------------------------------------