+++ examples/mkuvalid/mkuvalid.hs
@@ -5,31 +5,23 @@
 -- Copyright   : Copyright (c) 2019 Travis Cardwell
 -- License     : MIT
 --
--- 'TTC.valid' is used to create validated constants.  The sample credit card
--- is validated at compile-time.
+-- 'TTC.mkUntypedValid' is used to create @valid@ function, to validate
+-- constants of a specific type.  The sample username is validated at
+-- compile-time.
 ------------------------------------------------------------------------------
 
 {-# LANGUAGE TemplateHaskell #-}
 
 module Main (main) where
 
--- (ttc)
-import qualified Data.TTC as TTC
-
--- (ttc:example-valid)
-import CreditCard (CreditCard(CreditCard))
-
--- HLint does not support typed expression splices
-{-# ANN module "HLint: ignore" #-}
+-- (ttc-examples:example-valid)
+import qualified Username
+import Username (Username)
 
 ------------------------------------------------------------------------------
 
-sample :: CreditCard
-sample = CreditCard
-    $$(TTC.valid "John Q Doe")
-    $$(TTC.valid "1234 5678 9015")
-    $$(TTC.valid "2020-06")
-    $$(TTC.valid "123")
+sample :: Username
+sample = $(Username.valid "tcard")
 
 ------------------------------------------------------------------------------